open CCC from within the indicator

+++ setup.sh
@@ -10,7 +10,11 @@ else
 	cp amd-indicator /usr/bin/
 	chown root:root /usr/bin/amd-indicator
 	chmod 755 /usr/bin/amd-indicator
-	ln -s /usr/lib/i386-linux-gnu/mesa /usr/i386-linux-gnu/mesa
+	if ! [[ -L /usr/i386-linux-gnu/mesa ]]
+	then
+		ln -s /usr/lib/i386-linux-gnu/mesa /usr/i386-linux-gnu/mesa
+	fi
+	# ln -s /usr/lib/i386-linux-gnu/mesa /usr/i386-linux-gnu/mesa
 	cp 11switchable /etc/X11/Xsession.d/
 	chown root:root /etc/X11/Xsession.d/11switchable
 	chmod 644 /etc/X11/Xsession.d/11switchable
